+++ docs/changes.rst
@@ -10,6 +10,7 @@ Change Log
   * created a context manager chdir method, :meth:`.cd`
   * created :meth:`.put_r` to recursively put the contents of a local directory to a remote one
   * fixed a bug with :func:`.st_mode_to_int` on py3 (#52)
+  * :meth:`.listdir_attr` now returns a sorted list, sorted on filename
 
 * 0.2.7 (current, released 2014-05-24)